fix:빌터패턴,service 단일화
